Fix clippy warnings (#1589)
This fixes the following warnings:
- `clippy::format_in_format_args`
- `clippy::or_fun_call`
- `clippy::redundant_closure`
- `clippy::to_string_in_format_args`
- `clippy::unnecessary_cast`
- `unused_imports`

+++ src/builder/create_components.rs
@@ -203,7 +203,7 @@ impl CreateButton {
     }
 
     pub fn build(mut self) -> Value {
-        self.0.insert("type", from_number(2_u8 as u8));
+        self.0.insert("type", from_number(2_u8));
 
         utils::hashmap_to_json_map(self.0.clone()).into()
     }
